# outputs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/plans": {
      "post": {
        "description": "Crea el plan en estado Borrador con name obligatorio; description, startDate y endDate opcionales (D-3). CA-008-2 permite que el plan quede sin ninguna fase indefinidamente — RN-01 (estructura minima) y RN-02 (fecha de fin obligatoria) solo se verifican al publicar (HU-009), nunca al crear. Solo accesible para COACH (RN-18, D-12).",
        "operationId": "PlansController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plan creado exitosamente en estado Borrador",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Error de validacion en los datos del plan",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado — se requiere token JWT valido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado — se requiere rol Entrenador",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear el shell de un plan de entrenamiento",
        "tags": [
          "plans"
        ]
      },
      "get": {
        "description": "Lista todos los planes del entrenador autenticado, en cualquier estado (Borrador, Publicado, Finalizado, Archivado), paginado. `status` refleja el valor efectivo de cada plan (D-1). Solo accesible para COACH.",
        "operationId": "PlansController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Numero de pagina (base 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Cantidad de resultados por pagina (max 100)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de los planes propios del entrenador",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado — se requiere token JWT valido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado — se requiere rol Entrenador",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar los planes propios del entrenador autenticado",
        "tags": [
          "plans"
        ]
      }
    },
    "/api/plans/{id}": {
      "get": {
        "description": "Devuelve el plan con toda su estructura embebida: fases, semanas, sesiones y ejercicios de sesion, cada ejercicio con los datos de la version que la sesion conserva (RN-07). Lectura agregada en un solo GET (D-8). Solo el entrenador propietario del plan puede consultarlo (CA-008-6).",
        "operationId": "PlansController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del plan con su arbol completo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado — se requiere token JWT valido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "403": {
            "description": "No tiene acceso a este plan — solo el entrenador propietario puede consultarlo (CA-008-6)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Plan no encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener el detalle de un plan propio con su arbol completo",
        "tags": [
          "plans"
        ]
      },
      "patch": {
        "description": "name/description son editables en cualquier estado del plan (D-6). startDate/endDate solo se aceptan si el plan sigue en Borrador — si se envian con el plan ya publicado, rechaza con PLAN_IMMUTABLE (409) porque RN-31 inmoviliza la duracion total al publicar. Solo el entrenador propietario puede editarlo (CA-008-6).",
        "operationId": "PlansController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Error de validacion en los datos del plan",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado — se requiere token JWT valido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "403": {
            "description": "No tiene acceso a este plan — solo el entrenador propietario puede editarlo (CA-008-6)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Plan no encontrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "409": {
            "description": "El plan no permite modificar fechas fuera del estado Borrador (RN-31, D-6)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Editar los datos generales de un plan propio",
        "tags": [
          "plans"
        ]
      }
    },
...
    "/api/catalog/plans": {
      "get": {
        "description": "Lista los planes en estado Publicado y no vencidos (CA-011-1, CA-011-3). El filtro se aplica a nivel de query (status=PUBLISHED AND endDate>=now()), no sobre el campo computado — un plan Publicado ya vencido queda excluido de inmediato sin depender de que la columna persistida se haya actualizado a FINALIZED (D-1). Cada item incluye name, startDate, endDate y el perfil publico reducido del entrenador (id, name, avatarUrl). Accesible para cualquier usuario autenticado.",
        "operationId": "PlansCatalogController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Numero de pagina (base 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Cantidad de resultados por pagina (max 100)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de planes publicados y vigentes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanCatalogSummaryResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado — se requiere token JWT valido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar el catalogo de planes publicados",
        "tags": [
          "plans-catalog"
        ]
      }
    },
    "/api/catalog/plans/{id}": {
      "get": {
        "description": "Devuelve el detalle de un plan en estado Publicado y no vencido (CA-011-2): name, description, startDate, endDate, la estructura de fases/semanas/sesiones con SOLO la cantidad de ejercicios por sesion (no el detalle de cada uno), mas el perfil publico COMPLETO del entrenador (id, name, avatarUrl, description, bannerUrl — los dos ultimos con fallback null si el entrenador no tiene una CoachRequest aprobada, D-2b). Un plan que no esta Publicado-y-vigente responde ENTITY_NOT_FOUND (404), igual que si no existiera — el atleta no debe poder distinguir \"no existe\" de \"no esta publicado\". Accesible para cualquier usuario autenticado.",
        "operationId": "PlansCatalogController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del plan publicado, con el perfil publico completo del entrenador",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanCatalogDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado — se requiere token JWT valido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Plan no encontrado, o no esta en estado Publicado y vigente (mismo mensaje en ambos casos)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener el detalle de un plan publicado del catalogo",
        "tags": [
          "plans-catalog"
        ]
      }
    },
...
    "/api/instances": {
      "get": {
        "description": "Lista, paginada, todas las instancias del atleta autenticado — un item por plan al que llego a tener una suscripcion Activa (CA-023-4, CA-023-5). Resumen liviano (id, planId, planName, createdAt), sin el arbol completo — mismo criterio D-8 de EPICA-03. planName resuelto via join de solo lectura contra Plan. Solo accesible para ATHLETE.",
        "operationId": "InstancesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Numero de pagina (base 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Cantidad de resultados por pagina (max 100)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de las instancias propias del atleta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstanceSummaryResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado — se requiere token JWT valido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado — se requiere rol Atleta",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar las instancias personalizadas propias del atleta autenticado",
        "tags": [
          "instances"
        ]
      }
    },
...
  },
  "info": {
    "title": "fitmess API",
    "description": "API de la plataforma de fitness fitmess",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePlanDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del plan de entrenamiento",
            "example": "Fuerza e hipertrofia — bloque 1",
            "minLength": 2,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "description": "Descripcion del plan visible para el atleta en el catalogo (HU-011). Opcional al crear — se puede completar despues, incluso tras publicar (D-6).",
            "example": "Bloque de 8 semanas enfocado en fuerza maxima y volumen de hipertrofia para tren inferior.",
            "maxLength": 1000
          },
          "startDate": {
            "type": "string",
            "description": "Fecha de inicio del plan (YYYY-MM-DD). Opcional al crear — RN-02 solo la exige para PUBLICAR, nunca para guardar en Borrador (CA-008-2).",
            "example": "2026-09-01",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "description": "Fecha de fin del plan (YYYY-MM-DD). Opcional al crear, obligatoria para publicar (RN-02) — no existen planes open-ended.",
            "example": "2026-10-27",
            "format": "date"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "PlanCatalogSummaryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "UUID del plan",
            "example": "f6a7b8c9-d0e1-2345-fabc-456789012345",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Nombre del plan",
            "example": "Fuerza e hipertrofia — bloque 1"
          },
          "startDate": {
            "type": "string",
            "description": "Fecha de inicio del plan",
            "example": "2026-09-01",
            "format": "date",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "description": "Fecha de fin del plan",
            "example": "2026-10-27",
            "format": "date",
            "nullable": true
          },
          "coach": {
            "description": "Perfil publico reducido del entrenador que publico el plan",
            "allOf": [
              {
                "$ref": "#/components/schemas/CoachSummaryResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "coach"
        ]
      },
...
    }
  }
}